+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix crash in scheduler during DST transition times (#107)
 * Add read individual input command and optional publishing of individual input registers (#111)
 * [Internal Cleanup] Use signed integers for inverter registers/values (#115)
+* Add WriteParam functionality (lxp/cmd/all/set/param/X) (#117)
 
 
 # 0.9.0 - 2nd November 2022